+++ spec/unit/cli/env_spec.rb
@@ -19,6 +19,11 @@ RSpec.describe OneviewSDK::Cli do
       expect { command }.to output(/ONEVIEWSDK_PASSWORD\s+=\s'secret123'/).to_stdout_from_any_process
     end
 
+    it 'shows ONEVIEWSDK_DOMAIN' do
+      ENV['ONEVIEWSDK_DOMAIN'] = 'Other'
+      expect { command }.to output(/ONEVIEWSDK_DOMAIN\s+=\s'Other'/).to_stdout_from_any_process
+    end
+
     it 'shows ONEVIEWSDK_TOKEN' do
       expect { command }.to output(/ONEVIEWSDK_TOKEN\s+=\s'secret456'/).to_stdout_from_any_process
     end